Update INSTALL
If the user does not import the 2nd dvd, kickstart will fail to install losf because perl-IO-stringy is only available on dvd2. In this case the rest of the OS install continues. 

Worse, if you give the cobbler system profile a true repo_mirror of centos65 main repository (in addition to the unavoidable ks_mirror of centos65 from dvd), install will fail and halt with "perl-io-stringy not found" error message because it knows the package exists from the metadata it has from the centos65 main repository, but yum will fail to install it because of the apparently mandatory "priority 0" yum priority plugin setting of the ks_mirror centos65 dvd repository not allowing perl-io-stringy from the true repo_mirror of centos65 which has a higher priority of 99 by default.